Rename echoServer package to server
