+++ src/main/java/server/HTTPServerMain.java
@@ -1,11 +1,11 @@
-package echoServer;
+package server;
 
 import java.net.ServerSocket;
 
 /**
  * echoServer.Main runs the server.
  */
-public class Main {
+public class HTTPServerMain {
     public static void main(String[] args) throws Exception {
         try (ServerSocket serverSocket = new ServerSocket(3000)) {
             PortListenable portListener = new PortListener(serverSocket);
@@ -14,9 +14,9 @@ public class Main {
                     block.call();
                 }
             };
-            Application echoServer = new EchoServer();
+            Application httpServer = new HTTPServer();
 
-            new MyServer(infiniteLooper, portListener).serve(echoServer);
+            new MyServer(infiniteLooper, portListener).serve(httpServer);
         }
     }
 }
