+++ src/main/java/server/SocketReaderWriter.java
@@ -1,4 +1,4 @@
-package echoServer;
+package server;
 
 import java.io.BufferedReader;
 import java.io.IOException;
