+++ src/test/java/server/EchoServerTest.java
@@ -1,4 +1,4 @@
-package echoServer;
+package server;
 
 import org.junit.jupiter.api.Test;
